container/bin/fullget-gentooghpr.sh: rewrite
Github has started aggressively limiting frequent requests when you're not
logged in. This is an attempt to slow down fetching, while still keeping the
whole process as much automated as possible. This is only an issue if you have
a habit of launching multiple PR testers at once, or within short intervals.